Fixed compiling errors
iri2uri.py could not be compiled by py4a when buildozer was building the
application. This is similar to the problem addressed in
https://github.com/kivy/buildozer/issues/135. This has caused the first
few debug APKs to throw and error in py4a. The problem is caused by a
testing method within httplib2 that is using python special unicode
characters. This problem is addressed by commenting out test_uris() and
test_iri() and by removing the test folder. These files are not at all
used by main.py. Thus, it was safe to remove them. This solved
compilation errors and iri2uri.py was compiled into iri2uri.pyo