Added SearchBus Behaviour
The SearchBus class now provides the intended behaviour.
Upon searching for a bus stop number, returns all information about the
buses in operation.

If the user has not logged in, he will not be able to save his
preferences.
If the user is logged in, his preferences will be saved upon 'checking'
the checkbox and will be deleted upon 'unchecking' the checkbox.

All saved and deleted preferences are kept in a list which updates the
DreamFactory RESTful API real time via another thread