Added Facebook User information
The login behaviour now handles the retrieval of information from the
interaction between the android-facebook-sdk and the jnnius callbacks.

Upon clicking the login button, the user will be asked to give his
permissions for our Next Bus application.

I have also added for the app to allow for the kivy app to be paused and
resumed once the facebook window is closed after authentication.

The user's facebook id, firstname and lastname are sent to datadb.py to
instantiate a POST request to the REST API for mySQL storage.