Added self.busInstance and bug-fixing
self.busInstance was added so that we can have a ref to the GET request
itself to retrieve the busStopID and the ServiceNo in request
automatically. When the user searches for a bus stop, the bus classes
will reference to this to obtain the bus stop no and service no for
arrival timing retrieval dynamically.