Added python threading and functools modules
A separate thread will be used to send user updates to our REST API.

partial from functools allow us to bind arguments to our function when
adding event callbacks. This allows the checkbox to initiate the 'save'
activity upon 'checking'. This is because the callback does not allow
for any arguments by default.