Added search user input behaviour
A user can enter the bus stop number, or a string (the name of the bus
stop) This 2 methods will get the user's input and initiate the relevant
search. If the user enters a string, this starts a string search in the
CSV to give the user suggestions.

Otherwise, a search is initiated using the bus stop number